backend/pyproject.toml
[project]
name = "system-admin-api"
version = "0.1.0"
description = "FastAPI backend for system-admin"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.29.0",
    "aerich>=0.7.2",
    "fastapi[standard]>=0.115.0",
    "fastapi-utils>=0.8.0",
    "bcrypt>=4.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.20",
    "typing-inspect>=0.9.0",
    "tortoise-orm[asyncpg]>=0.21.0",
    "uvicorn[standard]>=0.32.0",
    "agno>=2.5.17",
    "openai>=2.32.0",
    "starlette-admin[tortoise]>=0.16.0",
]

[project.scripts]
api-shell = "app.scripts.shell:main"
db-migrate = "aerich:main"

[project.optional-dependencies]
dev = [
    "aiosqlite",
    "ipython>=8.18",
    "httpx",
    "pytest",
    "pytest-asyncio",
    "ruff>=0.15.11",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 88
src = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.aerich]
tortoise_orm = "app.core.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
